add api - get film by ID

+++ src/main/java/com/filmrental/FilmRental/repo/CustomerRepo.java
@@ -3,5 +3,5 @@ package com.filmrental.FilmRental.repo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.filmrental.FilmRental.model.Customer;
 
-public interface CustomerRepo extends JpaRepository<Customer, Long> {
+public interface CustomerRepo extends JpaRepository<Customer, Integer> {
 }
