+++ src/main/java/com/filmrental/FilmRental/repo/RentalRepository.java
@@ -3,5 +3,5 @@ package com.filmrental.FilmRental.repo;
 import com.filmrental.FilmRental.model.Rental;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RentalRepository extends JpaRepository<Rental, Long> {
+public interface RentalRepository extends JpaRepository<Rental, Integer> {
 }
